refactored create ingredient page
