+++ client/src/Components/IngredientEditor/index.js
@@ -1,9 +1,5 @@
-import React, { useState } from 'react';
-import { useHistory } from "react-router-dom";
-import Axios from 'axios';
+import React from 'react';
 import styled from 'styled-components';
-import { EditorState, convertToRaw } from 'draft-js';
-import * as utils from '../../Utils/user';
 import TextEditor from '../TextEditor';
 import ImageSelector from '../ImageSelector';
 import Input from '../RegisterForm/Input';
@@ -27,63 +23,27 @@ const ErrorMessage = styled.div`
 `;
 
 
-const IngredientEditor = () => {
-
-    const [title, setTitle] = useState('');
-    const [imageUrl, setImageUrl] = useState('');
-    const [editorState, setEditorState] = useState(EditorState.createEmpty());
-    const [error, setError] = useState(false);
-    const [errorMessage, setErrorMessage] = useState('');
-
-    const history = useHistory();
-
-    const saveIngredient = (e) => {
-
-        e.preventDefault();
-
-        const authToken = utils.getCookieByName('auth-token');
-
-        const description = JSON.stringify(convertToRaw(editorState.getCurrentContent()));
-
-        const data = {
-            name: title,
-            imageUrl: imageUrl,
-            description: description
-        };
-
-        Axios('http://localhost:5000/api/ingredient', {
-            method: 'POST',
-            headers: {
-                'content-type': 'application/json',
-                'Authorization': authToken
-            }, data: data
-        }).then(() => {
-            history.push('/admin');
-        }).catch((err) => {
-            setError(true);
-            setErrorMessage('Something went wrong');
-        });
-    };
+const IngredientEditor = (props) => {
 
     return (
-        <Wrapper onSubmit={saveIngredient}>
+        <Wrapper onSubmit={props.action}>
             <h1>Create Ingredient</h1>
             <h3>Name</h3>
             <Input
-                value={title}
-                onChange={setTitle}
+                value={props.name}
+                onChange={props.setName}
             />
             <h3>Cover Image</h3>
             <ImageSelector
-                imageUrl={imageUrl}
-                setImageUrl={setImageUrl}
+                imageUrl={props.imageUrl}
+                setImageUrl={props.setImageUrl}
             />
             <h3>Description</h3>
             <TextEditor
-                editorState={editorState}
-                setEditorState={setEditorState}
+                editorState={props.editorState}
+                setEditorState={props.setEditorState}
             />
-            {error && <ErrorMessage>{errorMessage}</ErrorMessage>}
+            {props.error && <ErrorMessage>{props.errorMessage}</ErrorMessage>}
             <Submit label={'Save Ingredient'} />
         </Wrapper>
     );
